finish converting source code to typescript
