+++ src/capytable.css
@@ -480,122 +480,48 @@ table.capytable.display > tbody > tr.selected:hover > * {
   box-shadow: inset 0 0 0 9999px #0d6efd !important;
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 1) !important;
 }
-table.capytable.order-column > tbody tr > .sorting_1,
-table.capytable.order-column > tbody tr > .sorting_2,
-table.capytable.order-column > tbody tr > .sorting_3,
-table.capytable.display > tbody tr > .sorting_1,
-table.capytable.display > tbody tr > .sorting_2,
-table.capytable.display > tbody tr > .sorting_3 {
+table.capytable.order-column > tbody tr > .sorting,
+table.capytable.display > tbody tr > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.019);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-column-ordering), 0.019);
 }
-table.capytable.order-column > tbody tr.selected > .sorting_1,
-table.capytable.order-column > tbody tr.selected > .sorting_2,
-table.capytable.order-column > tbody tr.selected > .sorting_3,
-table.capytable.display > tbody tr.selected > .sorting_1,
-table.capytable.display > tbody tr.selected > .sorting_2,
-table.capytable.display > tbody tr.selected > .sorting_3 {
+table.capytable.order-column > tbody tr.selected > .sorting,
+table.capytable.display > tbody tr.selected > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.919);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.919);
 }
-table.capytable.display > tbody > tr:nth-child(odd) > .sorting_1,
-table.capytable.order-column.stripe > tbody > tr:nth-child(odd) > .sorting_1 {
+table.capytable.display > tbody > tr:nth-child(odd) > .sorting,
+table.capytable.order-column.stripe > tbody > tr:nth-child(odd) > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.054);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-column-ordering), 0.054);
 }
-table.capytable.display > tbody > tr:nth-child(odd) > .sorting_2,
-table.capytable.order-column.stripe > tbody > tr:nth-child(odd) > .sorting_2 {
-  box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.047);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-column-ordering), 0.047);
-}
-table.capytable.display > tbody > tr:nth-child(odd) > .sorting_3,
-table.capytable.order-column.stripe > tbody > tr:nth-child(odd) > .sorting_3 {
-  box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.039);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-column-ordering), 0.039);
-}
-table.capytable.display > tbody > tr:nth-child(odd).selected > .sorting_1,
+table.capytable.display > tbody > tr:nth-child(odd).selected > .sorting,
 table.capytable.order-column.stripe
   > tbody
   > tr:nth-child(odd).selected
-  > .sorting_1 {
+  > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.954);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.954);
 }
-table.capytable.display > tbody > tr:nth-child(odd).selected > .sorting_2,
-table.capytable.order-column.stripe
-  > tbody
-  > tr:nth-child(odd).selected
-  > .sorting_2 {
-  box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.947);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.947);
-}
-table.capytable.display > tbody > tr:nth-child(odd).selected > .sorting_3,
-table.capytable.order-column.stripe
-  > tbody
-  > tr:nth-child(odd).selected
-  > .sorting_3 {
-  box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.939);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.939);
-}
-table.capytable.display > tbody > tr.even > .sorting_1,
-table.capytable.order-column.stripe > tbody > tr.even > .sorting_1 {
+table.capytable.display > tbody > tr.even > .sorting,
+table.capytable.order-column.stripe > tbody > tr.even > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.019);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-column-ordering), 0.019);
 }
-table.capytable.display > tbody > tr.even > .sorting_2,
-table.capytable.order-column.stripe > tbody > tr.even > .sorting_2 {
-  box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.011);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-column-ordering), 0.011);
-}
-table.capytable.display > tbody > tr.even > .sorting_3,
-table.capytable.order-column.stripe > tbody > tr.even > .sorting_3 {
-  box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.003);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-column-ordering), 0.003);
-}
-table.capytable.display > tbody > tr.even.selected > .sorting_1,
-table.capytable.order-column.stripe > tbody > tr.even.selected > .sorting_1 {
+table.capytable.display > tbody > tr.even.selected > .sorting,
+table.capytable.order-column.stripe > tbody > tr.even.selected > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.919);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.919);
 }
-table.capytable.display > tbody > tr.even.selected > .sorting_2,
-table.capytable.order-column.stripe > tbody > tr.even.selected > .sorting_2 {
-  box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.911);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.911);
-}
-table.capytable.display > tbody > tr.even.selected > .sorting_3,
-table.capytable.order-column.stripe > tbody > tr.even.selected > .sorting_3 {
-  box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.903);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.903);
-}
-table.capytable.display tbody tr:hover > .sorting_1,
-table.capytable.order-column.hover tbody tr:hover > .sorting_1 {
+table.capytable.display tbody tr:hover > .sorting,
+table.capytable.order-column.hover tbody tr:hover > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.082);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-hover), 0.082);
 }
-table.capytable.display tbody tr:hover > .sorting_2,
-table.capytable.order-column.hover tbody tr:hover > .sorting_2 {
-  box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.074);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-hover), 0.074);
-}
-table.capytable.display tbody tr:hover > .sorting_3,
-table.capytable.order-column.hover tbody tr:hover > .sorting_3 {
-  box-shadow: inset 0 0 0 9999px rgba(0, 0, 0, 0.062);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-hover), 0.062);
-}
-table.capytable.display tbody tr:hover.selected > .sorting_1,
-table.capytable.order-column.hover tbody tr:hover.selected > .sorting_1 {
+table.capytable.display tbody tr:hover.selected > .sorting,
+table.capytable.order-column.hover tbody tr:hover.selected > .sorting {
   box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.982);
   box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.982);
-}
-table.capytable.display tbody tr:hover.selected > .sorting_2,
-table.capytable.order-column.hover tbody tr:hover.selected > .sorting_2 {
-  box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.974);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.974);
-}
-table.capytable.display tbody tr:hover.selected > .sorting_3,
-table.capytable.order-column.hover tbody tr:hover.selected > .sorting_3 {
-  box-shadow: inset 0 0 0 9999px rgba(13, 110, 253, 0.962);
-  box-shadow: inset 0 0 0 9999px rgba(var(--dt-row-selected), 0.962);
 }
 table.capytable.compact thead th,
 table.capytable.compact thead td,
