+++ src/models/options.ts
@@ -1,0 +1,5 @@
+export interface IOptions {
+  searching: boolean;
+  paging: boolean;
+  ordering: boolean;
+}
